JBRULES-2835: revert session id type, back from long to int

+++ jbpm-bpmn2/src/test/java/org/jbpm/bpmn2/SimpleBPMNProcessTest.java
@@ -1021,7 +1021,7 @@ public class SimpleBPMNProcessTest extends JbpmTestCase {
 	
 	private StatefulKnowledgeSession restoreSession(StatefulKnowledgeSession ksession) {
 //		return ksession;
-		long id = ksession.getId();
+		int id = ksession.getId();
 		KnowledgeBase kbase = ksession.getKnowledgeBase();
 		Environment env = ksession.getEnvironment();
 		KnowledgeSessionConfiguration config = ksession.getSessionConfiguration();
